+++ sdk/textanalytics/ai-text-analytics/rollup.base.config.js
@@ -61,7 +61,7 @@ export function browserConfig(test = false, production = false) {
   const baseConfig = {
     input: input,
     output: {
-      file: "browser/azure-ai-textanalytics.js",
+      file: "browser/azure-ai-text-analytics.js",
       format: "umd",
       name: "Azure.CognitiveServicesTextAnalytics",
       sourcemap: true,
@@ -117,7 +117,7 @@ export function browserConfig(test = false, production = false) {
     // applies to test code, which causes all tests to be removed by tree-shaking.
     baseConfig.treeshake = false;
   } else if (production) {
-    baseConfig.output.file = "browser/azure-ai-textanalytics.min.js";
+    baseConfig.output.file = "browser/azure-ai-text-analytics.min.js";
     baseConfig.plugins.push(terser());
   }